Add "Max Pressure" tool, Kn->pressure graph
This commit moves the function for going from kn to pressure from the motor class to the propellant class and uses it for a new graph in the propellant editor. It also adds a new tool that takes in a desired max pressure and changes the nozzle throat so the motor won't exceed it.